+++ Cargo.toml
@@ -28,6 +28,7 @@ members = [
     "crates/fluvio-smartmodule",
     "crates/fluvio-smartmodule-derive",
     "crates/fluvio-smartmodule-package",
+    "crates/fluvio-smartmodule-metadata",
     "crates/fluvio-smartengine",
     "crates/fluvio-socket",
     "crates/fluvio-spu",
