+++ crates/fluvio-controlplane-metadata/Cargo.toml
@@ -3,7 +3,7 @@ name = "fluvio-controlplane-metadata"
 edition = "2021"
 version = "0.18.1"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
-description = "Fluvio metadata"
+description = "Metadata definition for Fluvio control plane"
 repository = "https://github.com/infinyon/fluvio"
 license = "Apache-2.0"
