+++ crates/fluvio-controlplane/Cargo.toml
@@ -3,7 +3,7 @@ name = "fluvio-controlplane"
 version = "0.0.0"
 edition = "2021"
 license = "Apache-2.0"
-description = "Fluvio control plane API"
+description = "API for Fluvio Control Plane"
 authors = ["fluvio.io"]
 repository = "https://github.com/infinyon/fluvio"
 publish = false
