+++ crates/fluvio-smartengine/README.md
@@ -1,0 +1,1 @@
+Fluvio SmartModule execution engine
