+++ crates/fluvio-smartengine/src/lib.rs
@@ -1,2 +1,8 @@
-mod smartmodule;
-pub use smartmodule::*;
+mod transforms;
+pub use transforms::*;
+
+pub type WasmSlice = (i32, i32, u32);
+mod engine;
+pub use engine::*;
+
+mod memory;
