+++ crates/fluvio-smartengine/src/memory.rs
@@ -10,7 +10,7 @@ const MEMORY: &str = "memory";
 
 /// Copy a byte array into an instance's linear memory
 /// and return the offset relative to the module's memory.
-pub fn copy_memory_to_instance<T>(
+pub(crate) fn copy_memory_to_instance<T>(
     store: &mut Store<T>,
     instance: &Instance,
     bytes: &[u8],
