+++ crates/fluvio-smartengine/src/transforms/join.rs
@@ -6,11 +6,11 @@ use tracing::{debug, instrument};
 use wasmtime::{AsContextMut, Trap, TypedFunc};
 
 use dataplane::smartmodule::{
-    SmartModuleInput, SmartModuleExtraParams, SmartModuleOutput, SmartModuleInternalError,
+    SmartModuleInput, SmartModuleOutput, SmartModuleExtraParams, SmartModuleInternalError,
 };
 use crate::{
     WasmSlice,
-    smartmodule::{SmartModuleWithEngine, SmartModuleContext, SmartModuleInstance, error::Error},
+    {SmartModuleWithEngine, SmartModuleContext, SmartModuleInstance, error::Error},
 };
 
 const JOIN_FN_NAME: &str = "join";
@@ -18,7 +18,7 @@ type OldJoinFn = TypedFunc<(i32, i32), i32>;
 type JoinFn = TypedFunc<(i32, i32, u32), i32>;
 
 #[derive(Debug)]
-pub struct SmartModuleJoinStream {
+pub struct SmartModuleJoin {
     base: SmartModuleContext,
     join_fn: JoinFnKind,
 }
@@ -46,7 +46,7 @@ impl JoinFnKind {
     }
 }
 
-impl SmartModuleJoinStream {
+impl SmartModuleJoin {
     pub fn new(
         module: &SmartModuleWithEngine,
         params: SmartModuleExtraParams,
@@ -67,8 +67,8 @@ impl SmartModuleJoinStream {
     }
 }
 
-impl SmartModuleInstance for SmartModuleJoinStream {
-    #[instrument(skip(self, input), name = "JoinStream")]
+impl SmartModuleInstance for SmartModuleJoin {
+    #[instrument(skip(self, input), name = "Join")]
     fn process(&mut self, input: SmartModuleInput) -> Result<SmartModuleOutput> {
         let slice = self.base.write_input(&input)?;
         debug!(len = slice.1, "WASM SLICE");
