+++ crates/fluvio-smartengine/src/transforms/mod.rs
@@ -1,0 +1,11 @@
+pub mod filter;
+pub mod map;
+pub mod array_map;
+pub mod filter_map;
+pub mod aggregate;
+pub mod join;
+pub mod file_batch;
+pub mod join_stream;
+pub(crate) mod error;
+
+
