+++ crates/fluvio-smartmodule-metadata/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "fluvio-smartmodule-metadata"
+version = "0.0.1"
+edition = "2021"
+publish =  false
+license = "Apache-2.0"
+authors = ["Fluvio Contributors <team@fluvio.io>"]
+categories = ["wasm"]
+keywords = ["streaming", "stream", "wasm", "fluvio","hub"]
+repository = "https://github.com/infinyon/fluvio"
+description = "Metadata for Fluvio Smart Modules"
+
+
+[lib]
+crate-type = ['lib']
+
+
+[dependencies]
+serde = "1.0.0"
+toml = { version = "0.5.9", optional = true }
