+++ crates/fluvio-spu/src/smartengine/mod.rs
@@ -144,6 +144,7 @@ impl SmartModuleContext {
         // then get smartmodule context
         let payload = match invocation.wasm {
             SmartModuleInvocationWasm::Predefined(name) => {
+            
                 if let Some(smartmodule) = ctx.smartmodule_localstore().spec(&name) {
                     let wasm = SmartModuleWasmCompressed::Gzip(smartmodule.spec.wasm.payload);
                     LegacySmartModulePayload {
