+++ crates/fluvio/Cargo.toml
@@ -67,7 +67,7 @@ fluvio-compression = { version = "0.2", path = "../fluvio-compression" }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 dirs = "4.0.0"
-fluvio-smartengine = { path = "../fluvio-smartengine/", optional = true, version = "0.3.0" }
+fluvio-smartengine = { path = "../fluvio-smartengine/", optional = true, version = "0.4.0" }
 
 [target.'cfg(unix)'.dependencies]
 fluvio-spu-schema = { version = "0.10.0", path = "../fluvio-spu-schema", features = [
